# Better native.Process.stop performance for linux backend